DEV: check old IPs with Kolide to determine onboarded devices.
Previously, we only checked the current IP address of the user to determine whether a device is onboarded or not. It creates problem when the IP address is not updated either in Kolide system or in Discourse. So, now we're including the old IP addresses from the `UserAuthTokenLog` table. Also, we're now creating a cookie to prevent repeated checking.