DEV: check old IPs with Kolide to determine onboarded devices. (#74)
Previously, we only checked the current IP address of the user to determine whether a device was onboarded or not. It creates a problem when the IP address is not updated in the Kolide system or Discourse. So, now we're including the old IP addresses from the `UserAuthTokenLog` table. Also, we're now creating a cookie to prevent repeated checking.