FEATURE: new checkbox to mark a device as mobile in onboarding (#89)
Sometimes we're unable to detect the mobile devices from their user agent to skip the Kolide onboarding. For example, in the latest iPadOS 13 Beta, Apple claimed desktop-class browsing in Safari. It seems Safari mimics macOS behavior and user agent. This commit gives the ability to users to choose whether the current device is mobile or not.

<img width="422" alt="Screenshot 2024-01-28 at 1 50 04 PM" src="https://github.com/discourse/discourse-kolide/assets/9372109/2e9edc9e-4925-4eb5-a857-44992f40fa15">